Move _configure_cmake and build to conanfile.py
